organize file structure, complete iteration zero
