+++ src/components/App.js
@@ -1,5 +1,5 @@
 import React, { Component } from 'react';
-import './App.css';
+import '../styles/App.css';
 
 class App extends Component {
   render() {
