+++ src/index.js
@@ -2,6 +2,6 @@ import React from 'react';
 import { render } from 'react-dom';
 import './index.css';
 
-import App from './App';
+import App from './components/App';
 
 render( <App />, document.getElementById('root') );
