+++ src/test/unit/iteration-0.test.js
@@ -1,4 +1,4 @@
-import DistrictRepository from '../../src/helper.js';
+import DistrictRepository from '../../helpers/helper.js';
 import kinderData from '../../data/kindergartners_in_full_day_program.js';
 
 describe('DistrictRepository iteration 0', () =>  {
@@ -12,6 +12,7 @@ describe('DistrictRepository iteration 0', () =>  {
   test('data coming in has no duplicates', () => {
     // uncomment out the tests that best fits your model
     // expect(district.stats.length).toBe(181);
+    console.log(district.stats)
     expect(Object.keys(district.stats).length).toBe(181);
   });
